Move to BigDecimal for currency values

+++ app/src/main/java/com/mmutert/trackmydebt/data/Transaction.kt
@@ -29,7 +29,7 @@ data class Transaction(
 
     val received: Boolean,
 
-    val amount: Long,
+    val amount: BigDecimal,
 
     val date: LocalDateTime,
